Updated UI elements in _Layout.cshtml and _BookListPartial.cshtml
Updated several UI elements across _Layout.cshtml and _BookListPartial.cshtml.
In _BookListPartial.cshtml, the hyperlink for the book title has been updated for a cleaner look.
In _Layout.cshtml, a placeholder for future SVG elements has been added, a new navigation item for the user profile has been added to the navbar, and the `pb-3` class has been moved from the `main` element to the `body-footer-divider` class, and the text color of the footer links has been changed to the default color.