Added padding, margin and new navigation item
Added padding-bottom to book cover image, margin-top to site content, and a new navigation item "My books".
In `_BookListPartial.cshtml`, pb-3 was added to the `img` tag to create some space below the book cover image.
In `_Layout.cshtml`, a `margin-top` of `60px` was added to the `site-wrapper` div to create some space at the top of the site content.
Also in `_Layout.cshtml`, a new navigation item "My books" was added to the navigation bar to allow users to navigate to a page that displays their books.